price input sanitation

+++ wsd2017gamestore/developer/forms.py
@@ -3,7 +3,7 @@ from store.models import Category
 
 class NewGameForm(forms.Form):
     game_name = forms.CharField(label='Game name', max_length=200)
-    game_price = forms.FloatField(label='Game price', max_value=99.99)
+    game_price = forms.DecimalField(label='Game price', decimal_places=2, max_value=99.99)
     game_url = forms.URLField(label='Game url')
     game_desc = forms.CharField(label='Game description', max_length=1000)
